Add files via upload
This commit adds all the R files I have been working with thus far to the master branch.  It includes early testing data and programming for DRC and NIG.
The commit is from WHO Laptop to github.com.  I previously tried to do upload the files through a commit from git gui, but it did not show up on github.  This time I initialized through the command line on my system76, and am uploading files directly to github.com from the WHO laptop.